refactor: extract util functions
