+++ lib/autoConfig.js
@@ -2,68 +2,12 @@ const fs = require('fs');
 const path = require('path');
 const yaml = require('js-yaml');
 
-function isObject(item) {
-  return item && typeof item === 'object' && !Array.isArray(item);
-}
-
-/**
- * Check if the reference value is defined and has a value
- * @param obj
- * @returns {boolean|boolean}
- */
-function hasValue(obj) {
-  return obj !== null && obj !== undefined;
-}
-
-/**
- * Get the property from a object ignoring case
- * E.g.: obj.getProperty('HeLo') === obj.getProperty('hElO')
- */
-function getProperty(object, key) {
-  return object[
-    Object.keys(object).find(k => k.toLowerCase() === key.toLowerCase())
-  ];
-}
-
-/**
- * Set the value for an EXISTING property. It finds the property ignoring the case. If the following operations
- * run in sequence the second will override the value of the first:
- * obj.setProperty('HeLo', 'hi')
- * obj.setProperty('hElO', 'hello')
- *
- * It also casts the new value to have the same data type. If cast fails the operation also fails
- */
-function setProperty(object, key, systemVariable) {
-  const objectKey = Object.keys(object).find(
-    k => k.toLowerCase() === key.toLowerCase()
-  );
-  const value = process.env[systemVariable];
-  // const castedValue = cast(object[objectKey], value);
-
-  switch (typeof object[objectKey]) {
-    case 'boolean':
-      if (value.toLowerCase() !== 'true' && value.toLowerCase() !== 'false') {
-        throw new Error(
-          `${systemVariable} contains invalid data. true/false expected.`
-        );
-      }
-      // eslint-disable-next-line no-param-reassign
-      object[objectKey] = value.toLowerCase() === 'true';
-      break;
-    case 'number':
-      if (!Number(value)) {
-        throw new Error(
-          `${systemVariable} contains invalid data. Number expected.`
-        );
-      }
-      // eslint-disable-next-line no-param-reassign
-      object[objectKey] = Number(value);
-      break;
-    default:
-      // eslint-disable-next-line no-param-reassign
-      object[objectKey] = value;
-  }
-}
+const {
+  isObject,
+  hasValue,
+  getPropertyCaseInsensitive,
+  setPropertyCaseInsensitive,
+} = require('./utils');
 
 // Configuration object shared in the whole application
 let config;
@@ -113,22 +57,33 @@ function findPropertyToBeOverride(configObj, tokens) {
 
   // Cases
   // The root path is already the sysvar name
-  if (hasValue(getProperty(configObj, tokens.join('')))) {
+  if (hasValue(getPropertyCaseInsensitive(configObj, tokens.join('')))) {
     return { node: configObj, key: tokens.join('') };
   }
 
   // End of the path, the property is there and there is no one else to check
-  if (hasValue(getProperty(configObj, current)) && !hasValue(next)) {
+  if (
+    hasValue(getPropertyCaseInsensitive(configObj, current)) &&
+    !hasValue(next)
+  ) {
     return { node: configObj, key: current };
   }
 
   // Start/middle of the path. Has the property and the next, keep calling
   if (
-    hasValue(getProperty(configObj, current)) &&
-    hasValue(getProperty(getProperty(configObj, current), next))
+    hasValue(getPropertyCaseInsensitive(configObj, current)) &&
+    hasValue(
+      getPropertyCaseInsensitive(
+        getPropertyCaseInsensitive(configObj, current),
+        next
+      )
+    )
   ) {
     const [, ...rest] = tokens;
-    return findPropertyToBeOverride(getProperty(configObj, current), rest);
+    return findPropertyToBeOverride(
+      getPropertyCaseInsensitive(configObj, current),
+      rest
+    );
   }
 
   // Find the property in this level but not in the next
@@ -136,11 +91,16 @@ function findPropertyToBeOverride(configObj, tokens) {
   // if it is the next
   // Covers: cookie.maxAge <-> COOKIE_MAX_AGE
   if (
-    hasValue(getProperty(configObj, current)) &&
-    !hasValue(getProperty(getProperty(configObj, current), next))
+    hasValue(getPropertyCaseInsensitive(configObj, current)) &&
+    !hasValue(
+      getPropertyCaseInsensitive(
+        getPropertyCaseInsensitive(configObj, current),
+        next
+      )
+    )
   ) {
     const [, ...rest] = tokens;
-    if (hasValue(getProperty(configObj, rest.join('')))) {
+    if (hasValue(getPropertyCaseInsensitive(configObj, rest.join('')))) {
       return { node: configObj, key: tokens.join('') };
     }
     return undefined;
@@ -164,7 +124,7 @@ function overrideConfigValuesFromSystemVariables(configObj) {
 
     const result = findPropertyToBeOverride(configObj, tokens);
     if (result && result.node && result.key) {
-      setProperty(result.node, result.key, systemVariable);
+      setPropertyCaseInsensitive(result.node, result.key, systemVariable);
     }
   });
 }
